fix(game): set claim_win_in_seconds integer type to u16 to avoid overflows under rare circumstances
(cherry picked from commit 69744f63e562364c1e62670f9e1c42042a19c2a1)

+++ src/models/game.rs
@@ -226,7 +226,7 @@ pub struct GameFull {
 pub struct OpponentGone {
     pub gone: bool,
     #[serde(rename = "claimWinInSeconds")]
-    pub claim_win_in_seconds: Option<u8>,
+    pub claim_win_in_seconds: Option<u16>,
 }
 
 #[skip_serializing_none]
